Split out the []byte-array generation for core-library namespaces
This eliminates the need to hand-maintain the declaration of the []byte arrays, allowing `gen_data.go` to generate those declarations and initialization (statically, in a func init(), or in a lazily-called function) as it sees fit.
