+++ core/gen_data/gen_data.go
@@ -15,9 +15,7 @@ var template string = `// Generated by gen_data. Don't modify manually!
 
 package core
 
-func init() {
-	{name}Data = []byte("{content}")
-}
+var {name}Data = []byte("{content}")
 `
 
 type FileInfo struct {
